+++ make/harbor.yml.tmpl
@@ -26,6 +26,9 @@ https:
 # Remember Change the admin password from UI after launching Harbor.
 harbor_admin_password: Harbor12345
 
+#TODO: remove this temporary flag before ships v1.11/v2, this should always be true
+registry_use_basic_auth: false
+
 # Harbor DB configuration
 database:
   # The password for the root user of Harbor DB. Change this before any production use.
