+++ make/photon/prepare/Dockerfile.base
@@ -1,5 +1,6 @@
 FROM photon:2.0
 
-RUN tdnf install -y python3 \
-    && tdnf install -y python3-pip
+RUN tdnf install -y python3 python3-pip httpd
 RUN pip3 install pipenv==2018.11.26
+
+
