+++ make/photon/prepare/templates/registry/config.yml.jinja
@@ -26,11 +26,17 @@ http:
   debug:
     addr: localhost:5001
 auth:
+{% if registry_use_basic_auth %}
+  htpasswd:
+    realm: harbor-registry-basic-realm
+    path: /etc/registry/passwd
+{% else %}
   token:
     issuer: harbor-token-issuer
     realm: {{public_url}}/service/token
     rootcertbundle: /etc/registry/root.crt
     service: harbor-registry
+{% endif %}
 validation:
   disabled: true
 notifications:
